+++ lib/e_menu/e_menu.tcl
@@ -1322,6 +1322,7 @@ proc ::em::callmenu {typ s1 {amp ""} {from ""}} {
   set geo [wm geometry .em]
   set geo [string range $geo [string first + $geo] end]
   # shift the new menu if it's shown above the current one
+  set geo [apave::checkGeometry $geo]
   if {$::em::solo && ($noME || $::em::ontop)} {
     lassign [split $geo +] -> x y
     set geo +[expr {20+$x}]+[expr {30+$y}]